feat: change work item data type

+++ internal/work/work_items.go
@@ -9,15 +9,15 @@ type WorkItem struct {
 	AccumulateGasLimit uint64             `json:"accumulate_gas_limit"`
 	ImportSegments     []ImportSegment    `json:"import_segments"`
 	Extrinsic          []ExtrinsicElement `json:"extrinsic"`
-	ExportCount        uint               `json:"export_count"` // number of export segments
+	ExportCount        uint16             `json:"export_count"` // number of export segments
 }
 
 type ImportSegment struct {
 	Hash  [32]byte `json:"tree_root"` // hash of segment root or work package
-	Index uint     `json:"index"`     // index of prior exported segments
+	Index uint16   `json:"index"`     // index of prior exported segments
 }
 
 type ExtrinsicElement struct {
 	Hash   [32]byte `json:"hash"`
-	Length uint     `json:"len"`
+	Length uint32   `json:"len"`
 }
